Crate pallet_nicks

Nicks Pallet

Config

Call

Overview

Nicks is an example pallet for keeping track of account names on-chain. It makes no effort to create a name hierarchy, be a DNS replacement or provide reverse lookups. Furthermore, the weights attached to this pallet’s dispatchable functions are for demonstration purposes only and have not been designed to be economically secure. Do not use this pallet as-is in production.

Interface

Dispatchable Functions

set_name - Set the associated name of an account; a small deposit is reserved if not already taken.

clear_name - Remove an account’s associated name; the deposit is returned.

kill_name - Forcibly remove the associated name; the deposit is lost.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.